Move editable fields components
